RDCC-62 Review commnets for Validator fixed
