+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/UpdateOrganisationRequestValidator.java
@@ -1,10 +1,15 @@
 package uk.gov.hmcts.reform.professionalapi.controller.request;
 
 import java.util.List;
+import java.util.UUID;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
+import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
+import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
 
 @Component
+@Slf4j
 public class UpdateOrganisationRequestValidator {
 
     private final List<UpdateOrganisationValidator> validators;
@@ -13,7 +18,19 @@ public class UpdateOrganisationRequestValidator {
         this.validators = validators;
     }
 
-    public void validate(OrganisationCreationRequest organisationCreationRequest, String organisationIdentifier) {
-        validators.forEach(v -> v.validate(organisationCreationRequest,organisationIdentifier));
+    public void validateStatus(Organisation existingOrganisation, OrganisationStatus inputStatus, UUID inputOrganisationIdentifier) {
+        validators.forEach(v -> v.validate(existingOrganisation, inputStatus, inputOrganisationIdentifier));
+    }
+
+    public UUID validateAndReturnInputOrganisationIdentifier(String inputOrganisationIdentifier) {
+        UUID orgIdentifier = null;
+        try {
+            orgIdentifier = UUID.fromString(inputOrganisationIdentifier);
+        } catch (IllegalArgumentException ex) {
+            String errorMessage = "Invalid organisationIdentifier provided organisationIdentifier: " + inputOrganisationIdentifier;
+            log.error(errorMessage);
+            throw new InvalidRequest(errorMessage);
+        }
+        return orgIdentifier;
     }
 }
