+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/UpdateOrganisationValidator.java
@@ -1,8 +1,11 @@
 package uk.gov.hmcts.reform.professionalapi.controller.request;
 
+import java.util.UUID;
+import uk.gov.hmcts.reform.professionalapi.domain.Organisation;
+import uk.gov.hmcts.reform.professionalapi.domain.OrganisationStatus;
 
 public interface UpdateOrganisationValidator {
 
-    void validate(OrganisationCreationRequest organisationCreationRequest, String inputOrganisationIdentifier);
+    void validate(Organisation existingOrganisation, OrganisationStatus inputStatus, UUID inputOrganisationIdentifier);
 
 }
